requirements.txt
aiofiles==25.1.0
chromadb==1.3.5
fastapi==0.122.0
llama_index==0.14.8
pdfplumber==0.11.8
protobuf==6.33.1
pydantic==2.12.5
pydantic_settings==2.12.0
pytest==9.0.1
python-dotenv==1.2.1
Requests==2.32.5
streamlit==1.51.0
uvicorn==0.38.0
google-genai== 1.52.0
python-multipart==0.0.20
qdrant-client==1.16.1
fastembed==0.7.3